extracted PreferenceReceiver logic into main split broadcast action names into packages prefixes
